+++ finagle-core/src/main/scala/com/twitter/finagle/pushsession/PushStackClient.scala
@@ -5,7 +5,7 @@ import com.twitter.finagle.client.{EndpointerModule, EndpointerStackClient}
 import com.twitter.finagle.context.Contexts
 import com.twitter.finagle.param.{Label, ProtocolLibrary}
 import com.twitter.util.Future
-import java.net.InetSocketAddress
+import java.net.SocketAddress
 
 /**
  * Base type for building a [[com.twitter.finagle.client.StackClient]] using
@@ -48,7 +48,7 @@ abstract class PushStackClient[Req, Rep, This <: PushStackClient[Req, Rep, This]
   /**
    * Construct a new [[PushTransporter]] with the appropriately configured pipeline.
    */
-  protected def newPushTransporter(ia: InetSocketAddress): PushTransporter[In, Out]
+  protected def newPushTransporter(sa: SocketAddress): PushTransporter[In, Out]
 
   /**
    * Construct a new push session from the provided [[PushChannelHandle]] generated
@@ -64,9 +64,9 @@ abstract class PushStackClient[Req, Rep, This <: PushStackClient[Req, Rep, This]
   final protected def endpointer: Stackable[ServiceFactory[Req, Rep]] =
     new EndpointerModule[Req, Rep](
       Seq(implicitly[Stack.Param[ProtocolLibrary]], implicitly[Stack.Param[Label]]), {
-        (prms: Stack.Params, ia: InetSocketAddress) =>
+        (prms: Stack.Params, sa: SocketAddress) =>
           val endpointClient = self.copy1(params = prms)
-          val transporter = endpointClient.newPushTransporter(ia)
+          val transporter = endpointClient.newPushTransporter(sa)
           // This assumes that the `toString` of the implementation is sufficiently descriptive.
           // Note: this should be kept in sync with the equivalent `StdStackClient` logic.
           endpointClient.registerTransporter(transporter.toString)