+++ finagle-http/src/main/scala/com/twitter/finagle/Http.scala
@@ -21,7 +21,7 @@ import com.twitter.finagle.toggle.Toggle
 import com.twitter.finagle.tracing._
 import com.twitter.finagle.transport.{Transport, TransportContext}
 import com.twitter.util.{Duration, Future, Monitor, StorageUnit, Time}
-import java.net.{InetSocketAddress, SocketAddress}
+import java.net.SocketAddress
 
 /**
  * A rich HTTP/1.1 client with a *very* basic URL fetcher. (It does not handle
@@ -236,7 +236,7 @@ object Http extends Client[Request, Response] with HttpRichClient with Server[Re
     protected def endpointer: Stackable[ServiceFactory[Request, Response]] = {
       new EndpointerModule[Request, Response](
         Seq(implicitly[Stack.Param[HttpImpl]], implicitly[Stack.Param[param.Stats]]), {
-          (prms: Stack.Params, addr: InetSocketAddress) =>
+          (prms: Stack.Params, addr: SocketAddress) =>
             val transporter = params[HttpImpl].transporter(prms)(addr)
             new ServiceFactory[Request, Response] {
               def apply(conn: ClientConnection): Future[Service[Request, Response]] =